+++ modules/matio/src/include/WhosMatioFile.hpp
@@ -16,28 +16,17 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "NelsonGateway.hpp"
-#include "loadmatBuiltin.hpp"
-#include "savematBuiltin.hpp"
-#include "ismatfileBuiltin.hpp"
-#include "whosmatBuiltin.hpp"
+#pragma once
 //=============================================================================
-using namespace Nelson;
+#include "nlsMatio_exports.h"
+#include "ArrayOf.hpp"
+#include "Evaluator.hpp"
+#include "Interface.hpp"
 //=============================================================================
-const std::wstring gatewayName = L"matio";
+namespace Nelson {
 //=============================================================================
-static const nlsGateway gateway[] = {
-    { "loadmat", Nelson::MatioGateway::loadmatBuiltin, 1, 1, CPP_BUILTIN },
-    { "savemat", Nelson::MatioGateway::savematBuiltin, 0, 1, CPP_BUILTIN },
-    { "ismatfile", Nelson::MatioGateway::ismatfileBuiltin, 1, 1, CPP_BUILTIN },
-    { "whosmat", Nelson::MatioGateway::whosmatBuiltin, 1, 1, CPP_BUILTIN },
-};
+NLSMATIO_IMPEXP ArrayOf
+WhosMatioFile(Interface *io, const std::wstring& filename, wstringVector names, bool asStruct);
 //=============================================================================
-NLSGATEWAYFUNC(gateway)
-//=============================================================================
-NLSGATEWAYINFO(gateway)
-//=============================================================================
-NLSGATEWAYREMOVE(gateway)
-//=============================================================================
-NLSGATEWAYNAME()
+} // namespace Nelson
 //=============================================================================